Have new reader use the input file set generated in the initialize step (#1129)
* Forming the names of the files in the input file set, storing them in
ReaderPoolBase::newInputFileName_ and using these names for creating the files in
the initialize step and reading the files in the load step.

* Fixed coding norm error.

* Switched ReaderSinglePool::load() function over to use the generated input files instead
of the original input file. Not quite working yet.

* WIP: Adjustment of distribution map is working, still have to stop using the pre-loaded
dateTime, etc. values.

* New reader is working, but a new issue surfaced with dimension variable data not appearing
in the correct order when parallel hdf5 is used in the writer. However, writing multiple files
works correctly.

* Fixed failing ctests. All tests pass and the new reader is producing the same results as
the old reader

* Added MPI barrier between the readerPool->initialize() and readerPool->load() steps. This
was put in to make sure any files opened/created in initialize are closed before proceeding
to load.

* Fixed reviewer comments: typos, added const where applicable.

* Fixed setNewInputFileName function to change the file suffix to ".nc4" on the new input
file set files. Note the input file set is always hdf5 for now.